sensitivity as double, parse input units

+++ GlobalQuakeCore/src/main/java/globalquake/core/database/Channel.java
@@ -11,7 +11,7 @@ public final class Channel implements Serializable {
     private static final long serialVersionUID = 6513039511077454262L;
     private final String code;
     private final String locationCode;
-    private long sensitivity;
+    private double sensitivity;
     private double sampleRate;
     private double latitude;
     private double longitude;
@@ -30,7 +30,7 @@ public final class Channel implements Serializable {
     }
 
     public Channel(String code, String locationCode, double sampleRate, double latitude, double longitude,
-                   double elevation, StationSource stationSource, long sensitivity) {
+                   double elevation, StationSource stationSource, double sensitivity) {
         this.code = code;
         this.locationCode = locationCode;
         this.sampleRate = sampleRate;
@@ -118,11 +118,11 @@ public final class Channel implements Serializable {
                 seedlinkNetwork -> seedlinkNetwork.selectedStations)).orElse(null);
     }
 
-    public long getSensitivity() {
+    public double getSensitivity() {
         return sensitivity;
     }
 
-    public void setSensitivity(long sensitivity) {
+    public void setSensitivity(double sensitivity) {
         this.sensitivity = sensitivity;
     }
 }
