+++ GlobalQuakeCore/src/main/java/globalquake/core/database/StationDatabase.java
@@ -157,7 +157,7 @@ public class StationDatabase implements Serializable {
 
     @SuppressWarnings("UnusedReturnValue")
     public static Channel getOrCreateChannel(Station station, String channelCode, String locationCode,
-                                             double lat, double lon, double alt, double sampleRate, StationSource stationSource, long sensitivity) {
+                                             double lat, double lon, double alt, double sampleRate, StationSource stationSource, double sensitivity) {
         Channel channel = getChannel(station, channelCode, locationCode);
         if(channel != null){
             if(channel.getSensitivity() <= 0 && sensitivity > 0){
