+++ GlobalQuakeCore/src/main/java/globalquake/core/station/GlobalStation.java
@@ -25,7 +25,7 @@ public class GlobalStation extends AbstractStation {
 
 	public GlobalStation(String networkCode, String stationCode, String channelName,
                          String locationCode, double lat, double lon, double alt,
-                         int id, SeedlinkNetwork seedlinkNetwork, long sensitivity) {
+                         int id, SeedlinkNetwork seedlinkNetwork, double sensitivity) {
 		super(networkCode, stationCode, channelName, locationCode, lat, lon, alt, id, seedlinkNetwork, sensitivity);
 		this.records = new TreeSet<>(Comparator.comparing(dataRecord -> dataRecord.getStartBtime().toInstant().toEpochMilli()));
